Initial refactor proposal
