+++ libcloudforensics/providers/aws/internal/account.py
@@ -19,25 +19,16 @@ analysis virtual machine to be used in incident response.
 """
 
 import binascii
-from datetime import datetime
 import json
-import logging
-import os
-import re
 
 import boto3
 import botocore
 
-log = logging.getLogger()
-
-EC2_SERVICE = 'ec2'
-CLOUDTRAIL_SERVICE = 'cloudtrail'
-ACCOUNT_SERVICE = 'sts'
-KMS_SERVICE = 'kms'
-# Default Amazon Machine Image to use for bootstrapping instances
-UBUNTU_1804_AMI = 'ami-0013b3aa57f8a4331'
-REGEX_TAG_VALUE = re.compile('^.{1,255}$')
-STARTUP_SCRIPT = 'scripts/startup.sh'
+from libcloudforensics.providers.aws.internal.common import EC2_SERVICE, \
+  REGEX_TAG_VALUE, GetTagForResourceType, GetInstanceTypeByCPU, \
+  ReadStartupScript, ACCOUNT_SERVICE, KMS_SERVICE
+from libcloudforensics.providers.aws.internal.ebs import AWSVolume
+from libcloudforensics.providers.aws.internal.ec2 import AWSInstance
 
 
 class AWSAccount:
@@ -492,8 +483,8 @@ class AWSAccount:
     except RuntimeError:
       pass
 
-    instance_type = self._GetInstanceTypeByCPU(cpu_cores)
-    startup_script = self._ReadStartupScript()
+    instance_type = GetInstanceTypeByCPU(cpu_cores)
+    startup_script = ReadStartupScript()
     if packages:
       startup_script = startup_script.replace('${packages[@]}', ' '.join(
           packages))
@@ -702,619 +693,3 @@ class AWSAccount:
     block_device_mapping = image['Images'][0]['BlockDeviceMappings'][0]
     block_device_mapping['Ebs']['VolumeSize'] = boot_volume_size
     return block_device_mapping
-
-  @staticmethod
-  def _GetInstanceTypeByCPU(cpu_cores):
-    """Return the instance type for the requested number of  CPU cores.
-
-    Args:
-      cpu_cores (int): The number of requested cores.
-
-    Returns:
-      str: The type of instance that matches the number of cores.
-
-    Raises:
-      ValueError: If the requested amount of cores is unavailable.
-    """
-
-    cpu_cores_to_instance_type = {
-        1: 't2.small',
-        2: 'm4.large',
-        4: 'm4.xlarge',
-        8: 'm4.2xlarge',
-        16: 'm4.4xlarge',
-        32: 'm5.8xlarge',
-        40: 'm4.10xlarge',
-        48: 'm5.12xlarge',
-        64: 'm4.16xlarge',
-        96: 'm5.24xlarge',
-        128: 'x1.32xlarge'
-    }
-    if cpu_cores not in cpu_cores_to_instance_type:
-      raise ValueError(
-          'Cannot start a machine with {0:d} CPU cores. CPU cores should be one'
-          ' of: {1:s}'.format(
-              cpu_cores, ', '.join(map(str, cpu_cores_to_instance_type.keys()))
-          ))
-    return cpu_cores_to_instance_type[cpu_cores]
-
-  @staticmethod
-  def _ReadStartupScript():
-    """Read and return the startup script that is to be run on the forensics VM.
-
-    Users can either write their own script to install custom packages,
-    or use the provided one. To use your own script, export a STARTUP_SCRIPT
-    environment variable with the absolute path to it:
-    "user@terminal:~$ export STARTUP_SCRIPT='absolute/path/script.sh'"
-
-    Returns:
-      str: The script to run.
-
-    Raises:
-      OSError: If the script cannot be opened, read or closed.
-    """
-
-    try:
-      startup_script = os.environ.get('STARTUP_SCRIPT')
-      if not startup_script:
-        # Use the provided script
-        startup_script = os.path.join(
-            os.path.dirname(os.path.realpath(__file__)), STARTUP_SCRIPT)
-      startup_script = open(startup_script)
-      script = startup_script.read()
-      startup_script.close()
-      return script
-    except OSError as exception:
-      raise OSError(
-          'Could not open/read/close the startup script {0:s}: {1:s}'.format(
-              startup_script, str(exception)))
-
-
-class AWSInstance:
-  """Class representing an AWS EC2 instance.
-
-  Attributes:
-    aws_account (AWSAccount): The account for the instance.
-    instance_id (str): The id of the instance.
-    region (str): The region the instance is in.
-    availability_zone (str): The zone within the region in which the instance
-        is.
-    name (str): The name tag of the instance, if existing.
-  """
-
-  def __init__(self,
-               aws_account,
-               instance_id,
-               region,
-               availability_zone,
-               name=None):
-    """Initialize the AWS EC2 instance.
-
-    Args:
-      aws_account (AWSAccount): The account for the instance.
-      instance_id (str): The id of the instance.
-      region (str): The region the instance is in.
-      availability_zone (str): The zone within the region in which the instance
-          is.
-      name (str): Optional. The name tag of the instance, if existing.
-    """
-
-    self.aws_account = aws_account
-    self.instance_id = instance_id
-    self.region = region
-    self.availability_zone = availability_zone
-    self.name = name
-
-  def GetBootVolume(self):
-    """Get the instance's boot volume.
-
-    Returns:
-      AWSVolume: Volume object if the volume is found.
-
-    Raises:
-      RuntimeError: If no boot volume could be found.
-    """
-
-    boot_device = self.aws_account.ResourceApi(
-        EC2_SERVICE).Instance(self.instance_id).root_device_name
-    volumes = self.ListVolumes()
-
-    for volume_id in volumes:
-      if volumes[volume_id].device_name == boot_device:
-        return volumes[volume_id]
-
-    error_msg = 'Boot volume not found for instance: {0:s}'.format(
-        self.instance_id)
-    raise RuntimeError(error_msg)
-
-  def GetVolume(self, volume_id):
-    """Get a volume attached to the instance by ID.
-
-    Args:
-      volume_id (str): The ID of the volume to get.
-
-    Returns:
-      AWSVolume: The AWSVolume object.
-
-    Raises:
-      RuntimeError: If volume_id is not found amongst the volumes attached
-          to the instance.
-    """
-
-    volume = self.ListVolumes().get(volume_id)
-    if not volume:
-      raise RuntimeError(
-          'Volume {0:s} is not attached to instance {1:s}'.format(
-              volume_id, self.instance_id))
-    return volume
-
-  def ListVolumes(self):
-    """List all volumes for the instance.
-
-    Returns:
-      dict: Dictionary mapping volumes to their respective AWSVolume object.
-    """
-
-    return self.aws_account.ListVolumes(
-        filters=[{
-            'Name': 'attachment.instance-id',
-            'Values': [self.instance_id]}])
-
-  def AttachVolume(self, volume, device_name):
-    """Attach a volume to the AWS instance.
-
-    Args:
-      volume (AWSVolume): The AWSVolume object to attach to the instance.
-      device_name (str): The device name for the volume (e.g. /dev/sdf).
-
-    Raises:
-      RuntimeError: If the volume could not be attached.
-    """
-
-    client = self.aws_account.ClientApi(EC2_SERVICE)
-    try:
-      client.attach_volume(
-          Device=device_name,
-          InstanceId=self.instance_id,
-          VolumeId=volume.volume_id)
-      volume.device_name = device_name
-    except client.exceptions.ClientError as exception:
-      raise RuntimeError('Could not attach volume {0:s}: {1:s}'.format(
-          volume.volume_id, str(exception)))
-
-
-class AWSElasticBlockStore:
-  """Class representing an AWS EBS resource.
-
-  Attributes:
-    aws_account (AWSAccount): The account for the resource.
-    region (str): The region the EBS is in.
-    availability_zone (str): The zone within the region in which the EBS is.
-    encrypted (bool): True if the EBS resource is encrypted, False otherwise.
-    name (str): The name tag of the EBS resource, if existing.
-  """
-
-  def __init__(self,
-               aws_account,
-               region,
-               availability_zone,
-               encrypted,
-               name=None):
-    """Initialize the AWS EBS resource.
-
-    Args:
-      aws_account (AWSAccount): The account for the resource.
-      region (str): The region the EBS is in.
-      availability_zone (str): The zone within the region in which the EBS is.
-      encrypted (bool): True if the EBS resource is encrypted, False otherwise.
-      name (str): Optional. The name tag of the EBS resource, if existing.
-    """
-
-    self.aws_account = aws_account
-    self.region = region
-    self.availability_zone = availability_zone
-    self.encrypted = encrypted
-    self.name = name
-
-
-class AWSVolume(AWSElasticBlockStore):
-  """Class representing an AWS EBS volume.
-
-  Attributes:
-    volume_id (str): The id of the volume.
-    aws_account (AWSAccount): The account for the volume.
-    region (str): The region the volume is in.
-    availability_zone (str): The zone within the region in which the volume is.
-    encrypted (bool): True if the volume is encrypted, False otherwise.
-    name (str): The name tag of the volume, if existing.
-    device_name (str): The device name (e.g. /dev/spf) of the
-        volume when it is attached to an instance, if applicable.
-  """
-
-  def __init__(self,
-               volume_id,
-               aws_account,
-               region,
-               availability_zone,
-               encrypted,
-               name=None,
-               device_name=None):
-    """Initialize an AWS EBS volume.
-
-    Args:
-      volume_id (str): The id of the volume.
-      aws_account (AWSAccount): The account for the volume.
-      region (str): The region the volume is in.
-      availability_zone (str): The zone within the region in which the volume
-          is.
-      encrypted (bool): True if the volume is encrypted, False otherwise.
-      name (str): Optional. The name tag of the volume, if existing.
-      device_name (str): Optional. The device name (e.g. /dev/spf) of the
-          volume when it is attached to an instance, if applicable.
-    """
-
-    super(AWSVolume, self).__init__(aws_account,
-                                    region,
-                                    availability_zone,
-                                    encrypted,
-                                    name)
-    self.volume_id = volume_id
-    self.device_name = device_name
-
-  def Snapshot(self, snapshot_name=None):
-    """Create a snapshot of the volume.
-
-    Args:
-      snapshot_name (str): Optional. Name tag of the snapshot.
-
-    Returns:
-      AWSSnapshot: A snapshot object.
-
-    Raises:
-      ValueError: If the snapshot name does not comply with the RegEx.
-      RuntimeError: If the snapshot could not be created.
-    """
-
-    timestamp = datetime.now().strftime('%Y%m%d%H%M%S')
-    if not snapshot_name:
-      snapshot_name = self.volume_id
-    truncate_at = 255 - len(timestamp) - 1
-    snapshot_name = '{0}-{1}'.format(snapshot_name[:truncate_at], timestamp)
-    if not REGEX_TAG_VALUE.match(snapshot_name):
-      raise ValueError('Snapshot name {0:s} does not comply with '
-                       '{1:s}'.format(snapshot_name, REGEX_TAG_VALUE.pattern))
-
-    client = self.aws_account.ClientApi(EC2_SERVICE)
-    try:
-      snapshot = client.create_snapshot(
-          VolumeId=self.volume_id,
-          TagSpecifications=[GetTagForResourceType('snapshot', snapshot_name)])
-
-      snapshot_id = snapshot.get('SnapshotId')
-      # Wait for snapshot completion
-      client.get_waiter('snapshot_completed').wait(SnapshotIds=[snapshot_id])
-    except (client.exceptions.ClientError,
-            botocore.exceptions.WaiterError) as exception:
-      raise RuntimeError('Could not create snapshot for volume {0:s}: '
-                         '{1:s}'.format(self.volume_id, str(exception)))
-
-    return AWSSnapshot(snapshot_id, self, name=snapshot_name)
-
-  def Delete(self):
-    """Delete a volume."""
-    client = self.aws_account.ClientApi(EC2_SERVICE)
-    try:
-      client.delete_volume(VolumeId=self.volume_id)
-    except client.exceptions.ClientError as exception:
-      raise RuntimeError('Could not delete volume {0:s}: {1:s}'.format(
-          self.volume_id, str(exception)))
-
-
-class AWSSnapshot(AWSElasticBlockStore):
-  """Class representing an AWS EBS snapshot.
-
-  Attributes:
-    snapshot_id (str): The id of the snapshot.
-    volume (AWSVolume): The volume from which the snapshot was taken.
-    name (str): The name tag of the snapshot, if existing.
-  """
-
-  def __init__(self, snapshot_id, volume, name=None):
-    """Initialize an AWS EBS snapshot.
-
-    Args:
-      snapshot_id (str): The id of the snapshot.
-      volume (AWSVolume): The volume from which the snapshot was taken.
-      name (str): Optional. The name tag of the snapshot, if existing.
-    """
-
-    super(AWSSnapshot, self).__init__(volume.aws_account,
-                                      volume.region,
-                                      volume.availability_zone,
-                                      volume.encrypted,
-                                      name)
-    self.snapshot_id = snapshot_id
-    self.volume = volume
-
-  def Delete(self):
-    """Delete a snapshot."""
-
-    client = self.aws_account.ClientApi(EC2_SERVICE)
-    try:
-      client.delete_snapshot(SnapshotId=self.snapshot_id)
-    except client.exceptions.ClientError as exception:
-      raise RuntimeError('Could not delete snapshot {0:s}: {1:s}'.format(
-          self.snapshot_id, str(exception)))
-
-  def ShareWithAWSAccount(self, aws_account_id):
-    """Share the snapshot with another AWS account ID.
-
-    Args:
-      aws_account_id (str): The AWS Account ID to share the snapshot with.
-    """
-
-    snapshot = self.aws_account.ResourceApi(EC2_SERVICE).Snapshot(
-        self.snapshot_id)
-    snapshot.modify_attribute(
-        Attribute='createVolumePermission',
-        CreateVolumePermission={
-            'Add': [{
-                'UserId': aws_account_id
-            }]
-        },
-        OperationType='add'
-    )
-
-
-class AWSCloudTrail:
-  """Class representing an AWS CloudTrail service.
-
-  Attributes:
-    aws_account (AWSAccount): The AWS account to use.
-  """
-
-  def __init__(self, aws_account):
-    """Initialize an AWS CloudTrail client.
-
-    Args:
-      aws_account (AWSAccount): The AWS account to use.
-    """
-
-    self.aws_account = aws_account
-
-  def LookupEvents(self,
-                   qfilter=(),
-                   starttime=None,
-                   endtime=None):
-    """Lookup events in the CloudTrail logs of this account.
-
-    Example usage:
-      # pylint: disable=line-too-long
-      # qfilter = 'key,value'
-      # starttime = datetime(2020,5,5,17,33,00)
-      # LookupEvents(qfilter=qfilter, starttime=starttime)
-      # Check documentation for qfilter details
-      # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/cloudtrail.html#CloudTrail.Client.lookup_events
-
-    Args:
-      qfilter (string): Optional. Filter for the query including 1 key and value.
-      starttime (datetime): Optional. Start datetime to add to query filter.
-      endtime (datetime): Optional. End datetime to add to query filter.
-
-    Returns:
-      list(dict): A list of events.
-    """
-
-    events = []
-
-    client = self.aws_account.ClientApi(CLOUDTRAIL_SERVICE)
-
-    params = {}
-    if qfilter:
-      k, v = qfilter.split(',')
-      qfilter = [{'AttributeKey': k, 'AttributeValue': v}]
-      params = {'LookupAttributes': qfilter}
-    if starttime:
-      params['StartTime'] = starttime
-    if endtime:
-      params['EndTime'] = endtime
-
-    while True:
-      response = client.lookup_events(**params)
-      for entry in response['Events']:
-        events.append(entry)
-      if 'NextToken' not in response:
-        return events
-      params['NextToken'] = response['NextToken']
-
-def CreateVolumeCopy(zone,
-                     instance_id=None,
-                     volume_id=None,
-                     src_account=None,
-                     dst_account=None):
-  """Create a copy of an AWS EBS Volume.
-
-  By default, the volume copy will be created in the same AWS account where
-  the source volume sits. If you want the volume copy to be created in a
-  different AWS account, you can specify one in the dst_account parameter.
-  The following example illustrates how you should configure your AWS
-  credentials file for such a use case.
-
-  # AWS credentials file
-  [default] # default account to use with AWS
-  aws_access_key_id=foo
-  aws_secret_access_key=bar
-
-  [investigation] # source account for a particular volume to be copied from
-  aws_access_key_id=foo1
-  aws_secret_access_key=bar1
-
-  [forensics] # destination account to create the volume copy in
-  aws_access_key_id=foo2
-  aws_secret_access_key=bar2
-
-  # Copies the boot volume from instance "instance_id" from the default AWS
-  # account to the default AWS account.
-  volume_copy = CreateVolumeCopy(zone, instance_id='instance_id')
-
-  # Copies the boot volume from instance "instance_id" from the default AWS
-  # account to the 'forensics' AWS account.
-  volume_copy = CreateVolumeCopy(
-      zone, instance_id='instance_id', dst_account='forensics')
-
-  # Copies the boot volume from instance "instance_id" from the
-  # 'investigation' AWS account to the 'forensics' AWS account.
-  volume_copy = CreateVolumeCopy(
-      zone,
-      instance_id='instance_id',
-      src_account='investigation',
-      dst_account='forensics')
-
-  Args:
-    zone (str): The zone within the region to create the new resource in.
-    instance_id (str): Optional. Instance ID of the instance using the volume
-        to be copied. If specified, the boot volume of the instance will be
-        copied. If volume_id is also specified, then the volume pointed by
-        that volume_id will be copied.
-    volume_id (str): Optional. ID of the volume to copy. If not set,
-        then instance_id needs to be set and the boot volume will be copied.
-    src_account (str): Optional. If the AWS account containing the volume
-        that needs to be copied is different from the default account specified
-        in the AWS credentials file, then you can specify it here (see
-        example above).
-    dst_account (str): Optional. If the volume copy needs to be created in a
-        different AWS account, you can specify it here (see example above).
-
-  Returns:
-    AWSVolume: An AWS EBS Volume object.
-
-  Raises:
-    RuntimeError: If there are errors copying the volume, or errors during
-        KMS key creation/sharing if the target volume is encrypted.
-    ValueError: If both instance_id and volume_id are missing.
-  """
-
-  if not instance_id and not volume_id:
-    raise ValueError('You must specify at least one of [instance_id, '
-                     'volume_id].')
-
-  source_account = AWSAccount(zone, aws_profile=src_account)
-  destination_account = AWSAccount(zone, aws_profile=dst_account)
-  kms_key_id = None
-
-  try:
-    if volume_id:
-      volume_to_copy = source_account.GetVolumeById(volume_id)
-    elif instance_id:
-      instance = source_account.GetInstanceById(instance_id)
-      volume_to_copy = instance.GetBootVolume()
-
-    log.info('Volume copy of {0:s} started...'.format(volume_to_copy.volume_id))
-    snapshot = volume_to_copy.Snapshot()
-
-    source_account_id = source_account.GetAccountInformation('Account')
-    destination_account_id = destination_account.GetAccountInformation(
-        'Account')
-
-    if source_account_id != destination_account_id:
-      if volume_to_copy.encrypted:
-        # Generate one-time use KMS key that will be shared with the
-        # destination account.
-        kms_key_id = source_account.CreateKMSKey()
-        source_account.ShareKMSKeyWithAWSAccount(
-            kms_key_id, destination_account_id)
-        temporary_volume = source_account.CreateVolumeFromSnapshot(
-            snapshot, kms_key_id=kms_key_id)
-        # The old snapshot is not needed anymore since we have created the
-        # temporary volume
-        snapshot.Delete()
-        # Get a new snapshot
-        snapshot = temporary_volume.Snapshot()
-        # Delete the temporary volume
-        temporary_volume.Delete()
-      snapshot.ShareWithAWSAccount(destination_account_id)
-
-    new_volume = destination_account.CreateVolumeFromSnapshot(
-        snapshot, volume_name_prefix='evidence')
-    snapshot.Delete()
-    # Delete the one-time use KMS key, if one was generated
-    source_account.DeleteKMSKey(kms_key_id)
-    log.info('Volume {0:s} successfully copied to {1:s}'.format(
-        volume_to_copy.volume_id, new_volume.volume_id))
-
-  except RuntimeError as exception:
-    error_msg = 'Copying volume {0:s}: {1!s}'.format(
-        (volume_id or instance_id), exception)
-    raise RuntimeError(error_msg)
-
-  return new_volume
-
-
-def StartAnalysisVm(vm_name,
-                    default_availability_zone,
-                    boot_volume_size,
-                    cpu_cores=4,
-                    ami=UBUNTU_1804_AMI,
-                    attach_volume=None,
-                    device_name=None,
-                    dst_account=None):
-  """Start a virtual machine for analysis purposes.
-
-  Look for an existing AWS instance with tag name vm_name. If found,
-  this instance will be started and used as analysis VM. If not found, then a
-  new vm with that name will be created, started and returned.
-
-  Args:
-    vm_name (str): The name for the virtual machine.
-    default_availability_zone (str): Default zone within the region to create
-        new resources in.
-    boot_volume_size (int): The size of the analysis VM boot volume (in GB).
-    cpu_cores (int): Optional. The number of CPU cores to create the machine
-        with. Default is 4.
-    ami (str): Optional. The Amazon Machine Image ID to use to create the VM.
-        Default is a version of Ubuntu 18.04.
-    attach_volume (AWSVolume): Optional. The volume to attach.
-    device_name (str): Optional. The name of the device (e.g. /dev/sdf) for the
-        volume to be attached. Mandatory if attach_volume is provided.
-    dst_account (str): Optional. The AWS account in which to create the
-        analysis VM. This is the profile name that is defined in your AWS
-        credentials file.
-
-  Returns:
-    tuple(AWSInstance, bool): a tuple with a virtual machine object
-        and a boolean indicating if the virtual machine was created or not.
-
-  Raises:
-    RuntimeError: If device_name is missing when attach_volume is provided.
-  """
-  aws_account = AWSAccount(default_availability_zone, aws_profile=dst_account)
-  analysis_vm, created = aws_account.GetOrCreateAnalysisVm(
-      vm_name, boot_volume_size, cpu_cores=cpu_cores, ami=ami)
-  if attach_volume:
-    if not device_name:
-      raise RuntimeError('If you want to attach a volume, you must also '
-                         'specify a device name for that volume.')
-    analysis_vm.AttachVolume(attach_volume, device_name)
-  return analysis_vm, created
-
-
-def GetTagForResourceType(resource, name):
-  """Create a dictionary for AWS Tag Specifications.
-
-  Args:
-    resource (str): The type of AWS resource.
-    name (str): The name of the resource.
-
-  Returns:
-    dict: A dictionary for AWS Tag Specifications.
-  """
-
-  return {
-      'ResourceType': resource,
-      'Tags': [
-          {
-              'Key': 'Name',
-              'Value': name
-          }
-      ]
-  }
