+++ tests/providers/startup_script.sh
